@@ -12,28 +12,28 @@
 		"preview": "vite preview",
 		"check": "svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
-		"lint": "prettier --check --plugin-search-dir=. . && eslint .",
-		"format": "prettier --write --plugin-search-dir=. ."
+		"lint": "prettier --check . && eslint .",
+		"format": "prettier --write ."
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/adapter-static": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/cookie": "^0.6.0",
-		"@types/eslint": "^8.56.7",
-		"eslint-plugin-svelte": "^2.36.0",
+		"@types/eslint": "^9.6.0",
+		"eslint-plugin-svelte": "^2.45.1",
 		"node-sass": "^9.0.0",
-		"prettier": "^3.2.5",
-		"prettier-plugin-svelte": "^3.2.3",
+		"prettier": "^3.3.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"sass": "^1.77.4",
-		"svelte": "^4.2.2",
-		"svelte-check": "^3.8.0",
-		"svelte-preprocess": "^5.1.4",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"svelte-preprocess": "^6.0.0",
 		"tslib": "^2.6.2",
-		"typescript": "^5.2.2",
-		"typescript-eslint": "^8.0.0-alpha.20",
-		"vite": "^5.0.3"
+		"typescript": "^5.5.0",
+		"typescript-eslint": "^8.0.0",
+		"vite": "^5.4.4"
 	},
 	"type": "module",
 	"dependencies": {
