@@ -2,8 +2,4 @@
 
 // See https://kit.svelte.dev/docs/types#the-app-namespace
 // for information about these interfaces
-declare namespace App {
-	interface Locals {
-		userid: string;
-	}
-}
+declare namespace App {}
