@@ -1,7 +1,7 @@
 export default class LoginResult {
-	public success: boolean;
-	public message: string;
-	public token: string;
-	public expiresIn: number;
-	public username: string;
+	public success: boolean = false;
+	public message: string = '';
+	public token: string = '';
+	public expiresIn: number = 0;
+	public username: string = '';
 }
