@@ -1,4 +1,4 @@
 export default class RegisterResult {
-	public success: boolean;
-	public message: string;
+	public success: boolean = false;
+	public message: string = '';
 }
