@@ -1,7 +1,7 @@
 import type ISmoke from './iSmoke';
 
 export default class SmokesSync {
-	public new: ISmoke[];
-	public updated: ISmoke[];
-	public deleted: string[];
+	public new: ISmoke[] = [];
+	public updated: ISmoke[] = [];
+	public deleted: string[] = [];
 }
