@@ -1,7 +1,7 @@
-import { writable } from 'svelte/store';
+import { writable, type Writable } from 'svelte/store';
 import TodaysSmokes from './models/todaysSmokes';
 
-export const online = writable(null);
+export const online: Writable<boolean | null> = writable(null);
 export const synced = writable(false);
 export const todaysSmokes = writable(new TodaysSmokes());
 export const statsDrawerIsOpen = writable(false);
